+++ .flake8
@@ -1,5 +1,5 @@
 [flake8]
-ignore = E266, E501, W503, F403, F401, W391, E203, E712, W293, F821
+ignore = E266, E501, W503, F403, F401, W391, E203, E712, W293, F821, W191
 max-line-length = 79
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
